Applied a simple web service(balance query) to the app
The web service is written in Java and run on the tomcat8 of Amazon
server(AWS). Since the IP address changed frequently so when we test
the app we need to keep the instance in AWS on, start the tomcat and
change the url(http://xxx.xxx.xxx.xxx:8080/…) of request in code to
match what shown on the server.